Move LineBBox to separate module and add lineleading configuration option
