+++ ts/output/common.ts
@@ -113,6 +113,7 @@ export abstract class CommonOutputJax<
     linebreaks: {                  // options for when overflow is linebreak
       inline: true,                   // true for browser-based breaking of inline equations
       width: '100%',                  // a fixed size or a percentage of the container width
+      lineleading: .2,                // the default lineleading in em units
       LinebreakVisitor: null,         // The LinebreakVisitor to use
     },
     wrapperFactory: null,          // The wrapper factory to use
