+++ ts/output/common/Wrapper.ts
@@ -34,7 +34,8 @@ import {CommonOutputJax} from '../common.js';
 import {CommonWrapperFactory} from './WrapperFactory.js';
 import {CommonMo} from './Wrappers/mo.js';
 import {CommonMrow} from './Wrappers/mrow.js';
-import {BBox, BBoxData} from '../../util/BBox.js';
+import {BBox} from '../../util/BBox.js';
+import {LineBBox} from './LineBBox.js';
 import {FontData, FontDataClass, DelimiterData,
         VariantData, CharData, CharOptions, DIRECTION, NOSTRETCH} from './FontData.js';
 
@@ -71,136 +72,6 @@ export type StyleData = {
     color: [string, string, string, string]
   }
 };
-
-/**
- * align and indent values
- */
-export type IndentData = [string, string];
-
-/**
- * A BBox with extra data about line breaking
- */
-export class LineBBox extends BBox {
-
-  /**
-   * The default line leading (for mo and mspace)
-   */
-  public static defaultLeading: number = .2;
-
-  /**
-   * Indentation data for the line break
-   */
-  public indentData: IndentData[] = null;
-
-  /**
-   * the lineleading for this break
-   */
-  public lineLeading: number;
-
-  /**
-   * The original BBox.L value
-   */
-  public originalL: number;
-
-  /**
-   * True if this bbox is at the beginning of a line
-   */
-  public isFirst: boolean = false;
-
-  /**
-   * The index of the child node containing this starting line break and the line within that child
-   */
-  public start: [number, number];
-  /**
-   * The index of the child node containingt the ending line break and the line within that child
-   */
-  public end: [number, number];
-
-  /**
-   * @param {BBox} bbox           The bbox to extend
-   * @param {number} leading      The lineleading value for the break
-   * @param {IndentData} indent   The align/shift information
-   * @return {ExtendedBBox}       The bbox extended
-   */
-  public static from(bbox: BBox, leading: number = LineBBox.defaultLeading, indent: IndentData[] = null): LineBBox {
-    const nbox = new this();
-    Object.assign(nbox, bbox);
-    nbox.lineLeading = leading;
-    if (indent) {
-      nbox.indentData = indent;
-    }
-    return nbox;
-  }
-
-  /**
-   * @override
-   */
-  constructor(def?: BBoxData, start: [number, number] = null) {
-    super(def);
-    this.originalL = this.L;
-    if (start) {
-      this.start = start;
-    }
-  }
-
-  /**
-   * @override
-   */
-  public append(cbox: LineBBox) {
-    if (this.isFirst) {
-      cbox.originalL = cbox.L;
-      cbox.L = 0;  // remove spacing after an operator with a linebreak after it
-    }
-    if (cbox.indentData) {
-      this.indentData = cbox.indentData;
-    }
-    this.lineLeading = cbox.lineLeading;
-    super.append(cbox);
-    this.isFirst = cbox.isFirst;
-  }
-
-  /**
-   * @override
-   */
-  public copy(): LineBBox {
-    const bbox = LineBBox.from(this);
-    bbox.indentData = this.indentData;
-    bbox.lineLeading = this.lineLeading;
-    return bbox;
-  }
-
-  /**
-   * @param {MmlNode} node    The MmlNode to get attributes from
-   */
-  public getIndentData(node: MmlNode) {
-    let {indentalign, indentshift, indentalignfirst, indentshiftfirst, indentalignlast, indentshiftlast} =
-      node.attributes.getAllAttributes() as StringMap;
-    if (indentalignfirst === 'indentalign') {
-      indentalignfirst = node.attributes.getInherited('indentalign') as string;
-    }
-    if (indentshiftfirst === 'indentshift') {
-      indentshiftfirst = node.attributes.getInherited('indentshift') as string;
-    }
-    if (indentalignlast === 'indentalign') {
-      indentalignlast = indentalign;
-    }
-    if (indentshiftlast === 'indentshift') {
-      indentshiftlast = indentshift;
-    }
-    this.indentData = [[indentalignfirst, indentshiftfirst],
-                       [indentalign, indentshift],
-                       [indentalignlast, indentshiftlast]] as IndentData[];
-  }
-
-  /**
-   * @param {LineBBox} bbox   The LineBBox whose indentData is to be copied
-   * @return {IndentData[]}   The copied array
-   */
-  protected copyIndentData(bbox: LineBBox): IndentData[] {
-    return bbox.indentData.map(([align, indent]) => [align, indent]);
-  }
-
-}
 
 /*********************************************************/
 /**
@@ -515,7 +386,14 @@ export class CommonWrapper<
    * Easy access to the linebreak visitor
    */
   get linebreaks() {
-    return this.factory.jax.linebreaks;
+    return this.jax.linebreaks;
+  }
+
+  /**
+   * Easy access to the linebreak options
+   */
+  get linebreakOptions() {
+    return this.jax.options.linebreaks;
   }
 
   /**
@@ -684,7 +562,7 @@ export class CommonWrapper<
           line.R = this.getBBox().R;
         }
       } else {
-        this.lineBBox[i] = LineBBox.from(this.getOuterBBox());
+        this.lineBBox[i] = LineBBox.from(this.getOuterBBox(), this.linebreakOptions.lineleading);
       }
     }
     return this.lineBBox[i];
